tests/requirements.txt
pytest==3.6.0
pytest-timeout==1.3.2
mock==2.0.0
freezegun==0.3.10
sqlalchemy==1.2.8
flask==0.12.2
Flask-Injector==0.10.1
flask-restplus==0.11.0
Flask-RESTful==0.3.6
psycopg2-binary==2.7.5
aws-requests-auth==0.4.1
boto3==1.7.36
docker-compose==1.17.1
flask-sqlalchemy==2.3.2
elasticsearch==6.2.0
click==6.7
flake8
warrant==0.6.1
-e functions
-e assets
